Improved efficiency
The Tweet-friendly bill name is now generated along with the URL name, so retrieveEdits.py no longer has to do it every time the same page is updated.

I also deleted the manualWhitelist stuff, since it's an unnecessary complication and can be handled more elegantly. Stay tuned on that front.